requirements.txt
fastapi[all]
uvicorn
psycopg2-binary
pandas
numpy
statsmodels
arch
Patsy
Scipy
yfinance
apscheduler